Move main files into subcommand folders
