+++ Commands/Administrator/AutoRole/AutoRole.js
@@ -1,7 +1,7 @@
 const Discord = require('discord.js');
 
-const AutoRoleSetup = require('./AutoRole/AutoRole.setup');
-const AutoRoleDisable = require('./AutoRole/AutoRole.disable');
+const AutoRoleSetup = require('./AutoRole.setup');
+const AutoRoleDisable = require('./AutoRole.disable');
 
 module.exports = {
     data: new Discord.SlashCommandBuilder()
