+++ Commands/Moderator/Warns/Warns.js
@@ -1,7 +1,7 @@
 const Discord = require('discord.js');
 
-const WarnsView = require('./Warns/Warns.view');
-const WarnsRemoveLog = require('./Warns/Warns.removelog');
+const WarnsView = require('./Warns.view');
+const WarnsRemoveLog = require('./Warns.removelog');
 
 module.exports = {
     data: new Discord.SlashCommandBuilder()
